requirements.txt
streamlit
pandas
sqlalchemy
streamlit-authenticator
plotly
pyyaml
apscheduler